refactor: move supported blocks to feature compatibilization
